requirements.txt
openai
langchain
langchain_community
flask
python-dotenv
streamlit
langchain_google_genai